Add CommentController
